Moved ReadDir and ReadDirRecursive in their own file
