+++ paths.go
@@ -306,47 +306,6 @@ func (p *Path) IsDirCheck() (bool, error) {
 	return false, err
 }
 
-// ReadDir returns a PathList containing the content of the directory
-// pointed by the current Path
-func (p *Path) ReadDir() (PathList, error) {
-	infos, err := ioutil.ReadDir(p.path)
-	if err != nil {
-		return nil, err
-	}
-	paths := PathList{}
-	for _, info := range infos {
-		path := p.Clone().Join(info.Name())
-		paths.Add(path)
-	}
-	return paths, nil
-}
-
-// ReadDirRecursive returns a PathList containing the content of the directory
-// and its subdirectories pointed by the current Path
-func (p *Path) ReadDirRecursive() (PathList, error) {
-	infos, err := ioutil.ReadDir(p.path)
-	if err != nil {
-		return nil, err
-	}
-	paths := PathList{}
-	for _, info := range infos {
-		path := p.Clone().Join(info.Name())
-		paths.Add(path)
-
-		if isDir, err := path.IsDirCheck(); err != nil {
-			return nil, err
-		} else if isDir {
-			subPaths, err := path.ReadDirRecursive()
-			if err != nil {
-				return nil, err
-			}
-			paths.AddAll(subPaths)
-		}
-
-	}
-	return paths, nil
-}
-
 // CopyTo copies the contents of the file named src to the file named
 // by dst. The file will be created if it does not already exist. If the
 // destination file exists, all it's contents will be replaced by the contents
